test: move concurrent navigation test into universal router suite
