+++ test/nuxt/universal/router.test.ts
@@ -13,4 +13,28 @@ describe('universal router', () => {
       redirectedFrom: undefined,
     })
   })
+
+  it('applies the latest navigation when concurrent navigations race', async () => {
+    const router = useRouter()
+
+    let resolveSlowGuard: () => void
+    let slowGuardEntered: (() => void) | undefined
+    const entered = new Promise<void>((resolve) => { slowGuardEntered = resolve })
+    const removeGuard = router.beforeEach(async (to) => {
+      if (to.path === '/slow') {
+        slowGuardEntered!()
+        await new Promise<void>((resolve) => { resolveSlowGuard = resolve })
+      }
+    })
+
+    const older = router.push('/slow')
+    await entered
+    const newer = router.push('/fast')
+    await newer
+    resolveSlowGuard!()
+    await older
+    removeGuard()
+
+    expect(router.currentRoute.value.path).toBe('/fast')
+  })
 })
